+++ saleor/graphql/warehouse/bulk_mutations/__init__.py
@@ -1,3 +1,3 @@
-from .stocks_bulk_update import StocksBulkUpdate
+from .stock_bulk_update import StockBulkUpdate
 
-__all__ = ["StocksBulkUpdate"]
+__all__ = ["StockBulkUpdate"]